в пути изменил v0 на v1, везде изменил нотацию на camelCase, убрал тавтологиии в ппути методов, убрал Optional в DataRepository, добавил обработку ошибок и добавил больше информативности, положил swaggerconfig в отдельную папку, в changelog-master использовую includeAll

+++ dictionary/src/main/java/neoflex/dictionary/controller/DataController.java
@@ -7,8 +7,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.AllArgsConstructor;
 import neoflex.dictionary.model.Data;
-import neoflex.dictionary.model.Dictionary;
 import neoflex.dictionary.service.DataService;
+import neoflex.dictionary.service.DictionaryService;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,10 +16,11 @@ import java.util.UUID;
 
 @AllArgsConstructor
 @RestController
-@RequestMapping("/api/v0/data")
+@RequestMapping("/api/v1/data")
 @Tag(name="DataController", description="Контролер для управлением Data")
 public class DataController {
 
+    private final DictionaryService dictionaryService;
     private DataService dataService;
 
     @Operation(
@@ -33,20 +34,26 @@ public class DataController {
     @Operation(
             summary = "Выводить все Data которые принадлежат Dictionary"
     )
-    @GetMapping("/{dictionary_id}")
-    public List<Data> findAllDataById(@PathVariable@Parameter(description = "Идентификатор Dictionary") UUID dictionary_id) {
-        return dataService.findAllDataByDictionary(dictionary_id);
+    @GetMapping("/{dictionaryId}")
+    public List<Data> findAllDataById(@PathVariable@Parameter(description = "Идентификатор Dictionary") UUID dictionaryId) {
+        return dataService.findAllDataByDictionary(dictionaryId);
     }
 
 
     @Operation(
             summary = "сохраняет Data",
             responses = {
-            @ApiResponse(description = "возвращает Data saved")
+            @ApiResponse(description = "возвращает Data saved или this dictionary does not exist если данного dictionary не существует")
             }
     )
-    @PostMapping("save_data")
-    public String saveData(@RequestBody Data data) {
+    @PostMapping("data")
+    public String saveData(@RequestBody(required = false) Data data) {
+        if (data == null) {
+            data = Data.defaultValue;
+        }
+        if (dictionaryService.findDictionary(data.getDictionary()) == null){
+            return "this dictionary does not exist";
+        }
         dataService.saveData(data);
         return "data saved";
     }
@@ -54,11 +61,20 @@ public class DataController {
     @Operation(
             summary = "удаляет Data",
             responses = {
-                    @ApiResponse(description = "возвращает Data deleted")
+                    @ApiResponse(description = "возвращает Data deleted или all data already deleted если при удалении по умолчанию нет ни одного data или data does not exist если данного data не существует")
             }
     )
-    @DeleteMapping("delete_data")
-    public String deleteData(@RequestBody Data data) {
+    @DeleteMapping("data")
+    public String deleteData(@RequestBody(required = false) Data data) {
+        if (data == null) {
+            data = dataService.findFirst();
+            if (data == null) {
+                return "all data already deleted";
+            }
+        }
+        if (dataService.findData(data) == null) {
+            return "data does not exist";
+        }
         dataService.deleteData(data);
         return "data deleted";
     }
